Break out pagination for public Nodes and quickfiles
